+++ cassandra/query.py
@@ -25,6 +25,7 @@ import struct
 import time
 import six
 from six.moves import range, zip
+import warnings
 
 from cassandra import ConsistencyLevel, OperationTimedOut
 from cassandra.util import unix_time_from_uuid1
@@ -83,6 +84,39 @@ def tuple_factory(colnames, rows):
     """
     return rows
 
+class PseudoNamedTupleRow(object):
+    """
+    Helper class for pseudo_named_tuple_factory. These objects provide an
+    __iter__ interface, as well as index- and attribute-based access to values,
+    but otherwise do not attempt to implement the full namedtuple or iterable
+    interface.
+    """
+    def __init__(self, ordered_dict):
+        self._dict = ordered_dict
+        self._tuple = tuple(ordered_dict.values())
+
+    def __getattr__(self, name):
+        return self._dict[name]
+
+    def __getitem__(self, idx):
+        return self._tuple[idx]
+
+    def __iter__(self):
+        return iter(self._tuple)
+
+    def __repr__(self):
+        return '{t}({od})'.format(t=self.__class__.__name__,
+                                  od=self._dict)
+
+
+def pseudo_namedtuple_factory(colnames, rows):
+    """
+    Returns each row as a :class:`.PseudoNamedTupleRow`. This is the fallback
+    factory for cases where :meth:`.named_tuple_factory` fails to create rows.
+    """
+    return [PseudoNamedTupleRow(od)
+            for od in ordered_dict_factory(colnames, rows)]
+
 
 def named_tuple_factory(colnames, rows):
     """
@@ -116,6 +150,20 @@ def named_tuple_factory(colnames, rows):
     clean_column_names = map(_clean_column_name, colnames)
     try:
         Row = namedtuple('Row', clean_column_names)
+    except SyntaxError:
+        warnings.warn(
+            "Failed creating namedtuple for a result because there were too "
+            "many columns. This is due to a Python limitation that affects "
+            "namedtuple in Python 3.0-3.6 (see issue18896). The row will be "
+            "created with {substitute_factory_name}, which lacks some namedtuple "
+            "features and is slower. To avoid slower performance accessing "
+            "values on row objects, Upgrade to Python 3.7, or use a different "
+            "row factory. (column names: {colnames})".format(
+                substitute_factory_name=pseudo_namedtuple_factory.__name__,
+                colnames=colnames
+            )
+        )
+        return pseudo_namedtuple_factory(colnames, rows)
     except Exception:
         clean_column_names = list(map(_clean_column_name, colnames))  # create list because py3 map object will be consumed by first attempt
         log.warning("Failed creating named tuple for results with column names %s (cleaned: %s) "